Allow to always preview whats new page
After initial notification there was no way to view whats new page again so we've added it to vim toolbar
